requirements.txt
albucore==0.0.23
albumentations==2.0.5
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.33.0
networkx==3.4.2
nibabel==5.3.2
numpy==2.1.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==76.1.0
SimpleITK==2.4.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.39
streamlit==1.44.1
stringzilla==3.12.3
svgwrite==1.4.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
toml==0.10.2

tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.100
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
